Module interledger::ildcp

Interledger Dynamic Configuration Protocol (ILDCP)

interledger-ildcp

Client and server implementations of the Interledger Dynamic Configuration Protocol (ILDCP).

This is used by clients to query for their ILP address and asset details such as asset code and scale.

Structs

IldcpRequest

IldcpResponse

IldcpResponseBuilder

IldcpService
A simple service that intercepts incoming ILDCP requests and responds using the information in the Account struct.

Traits

IldcpAccount

Functions

get_ildcp_info
Get the ILP address and asset details for a given account.

is_ildcp_request