Trait interledger::service::Account

pub trait Account: Send + Clone + Debug {
    type AccountId: Send + Copy + Sync + Eq + Default + FromStr + Hash + Debug + Display + Serialize;
    fn id(&self) -> Self::AccountId;

    fn username(&self) -> &Username;
}

The base trait that Account types from other Services extend. This trait only assumes that the account has an ID that can be compared with others.

Each service can extend the Account type to include additional details they require. Store implementations will implement these Account traits for a concrete type that they will load from the database.

Associated Types

type AccountId: Send + Copy + Sync + Eq + Default + FromStr + Hash + Debug + Display + Serialize

Loading content...
Required methods

fn id(&self) -> Self::AccountId

fn username(&self) -> &Username

Loading content...