Module interledger::spsp

Simple Payment Setup Protocol (SPSP) sender and query responder

interledger-spsp

Client and server implementations of the Simple Payment Setup Protocol (SPSP).

This uses a simple HTTPS request to establish a shared key between the sender and receiver that is used to authenticate ILP packets sent between them. SPSP uses the STREAM transport protocol for sending money and data over ILP.

Structs

SpspResponder
A Hyper::Service that responds to incoming SPSP Query requests with newly generated details for a STREAM connection.

SpspResponse

Enums

Error

Functions

pay
Query the details of the given Payment Pointer and send a payment using the STREAM protocol.

query