Struct interledger::store_memory::InMemoryStore

pub struct InMemoryStore { /* fields omitted */ }

A simple in-memory store intended primarily for testing and stateless sender/receiver services that are passed all of the relevant account details when the store is instantiated.

Methods

impl InMemoryStore

pub fn new<impl IntoIterator>(
accounts: impl IntoIterator
) -> InMemoryStore where
impl IntoIterator: IntoIterator<Item = AccountBuilder>,

pub fn default() -> InMemoryStore

pub fn from_accounts<impl IntoIterator>(
accounts: impl IntoIterator
) -> InMemoryStore where
impl IntoIterator: IntoIterator<Item = Account>,

pub fn add_account(&self, account: Account)