Module interledger::stream

STREAM Protocol sender and receiver

interledger-stream

Client and server implementations of the Interledger STREAM transport protocol.

STREAM is responsible for splitting larger payments and messages into smaller chunks of money and data, and sending them over ILP.

Structs

ConnectionGenerator
A STREAM connection generator that creates destination_account and shared_secret values based on a single root secret.

StreamReceiverService
An OutgoingService that fulfills incoming STREAM packets.

Enums

Error

Functions

send_money
Send a given amount of money using the STREAM transport protocol.